Separate observations parsing from validation
