+++ src/ert/config/parsing/__init__.py
@@ -9,7 +9,11 @@ from .forward_model_schema import init_forward_model_schema
 from .history_source import HistorySource
 from .hook_runtime import HookRuntime
 from .lark_parser import parse, parse_contents, read_file
-from .observations_parser import ObservationType
+from .observations_parser import (
+    ObservationConfigError,
+    ObservationType,
+    parse_observations,
+)
 from .queue_system import QueueSystem, QueueSystemWithGeneric
 from .schema_item_type import SchemaItemType
 from .workflow_job_keywords import WorkflowJobKeys
@@ -28,6 +32,7 @@ __all__ = [
     "ForwardModelStepKeys",
     "HistorySource",
     "HookRuntime",
+    "ObservationConfigError",
     "ObservationType",
     "QueueSystem",
     "QueueSystemWithGeneric",
@@ -41,5 +46,6 @@ __all__ = [
     "init_workflow_schema",
     "parse",
     "parse_contents",
+    "parse_observations",
     "read_file",
 ]
