+++ tests/ert/unit_tests/config/parsing/test_observations_parser.py
@@ -5,17 +5,18 @@ import hypothesis.extra.lark as stlark
 import pytest
 from hypothesis import given
 
-from ert.config.parsing.observations_parser import (
+from ert.config._observations_input import (
     GenObsValues,
     HistoryValues,
-    ObservationConfigError,
-    ObservationType,
     Segment,
     SummaryValues,
-    _parse_content_list,
-    _validate_conf_content,
+    validate_observations_input,
+)
+from ert.config.parsing.observations_parser import (
+    ObservationConfigError,
+    ObservationType,
     observations_parser,
-    parse_content,
+    parse_observations,
 )
 
 observation_contents = stlark.from_lark(observations_parser)
@@ -25,8 +26,8 @@ observation_contents = stlark.from_lark(observations_parser)
 @given(observation_contents)
 def test_parsing_contents_succeeds_or_gives_config_error(contents):
     with suppress(ObservationConfigError):
-        _ = _validate_conf_content(
-            ".", _parse_content_list(contents, "observations.txt")
+        _ = validate_observations_input(
+            ".", parse_observations(contents, "observations.txt")
         )
 
 
@@ -73,7 +74,7 @@ def file_contents():
 
 
 def test_parse(file_contents):
-    assert _parse_content_list(
+    assert parse_observations(
         file_contents,
         "",
     ) == [
@@ -124,11 +125,11 @@ def test_that_unexpected_character_gives_observation_config_error():
         ObservationConfigError,
         match=r"Line 1.*include a;",
     ):
-        _parse_content_list(content="include a;", filename="")
+        parse_observations(content="include a;", filename="")
 
 
 def test_that_double_comments_are_handled():
-    assert _parse_content_list(
+    assert parse_observations(
         """
             SUMMARY_OBSERVATION -- foo -- bar -- baz
                         FOPR;
@@ -142,17 +143,17 @@ def test_validate(file_contents):
     Path("wpr_diff_idx.txt").write_text("", encoding="utf8")
     Path("wpr_diff_obs.txt").write_text("", encoding="utf8")
     print(
-        _validate_conf_content(
+        validate_observations_input(
             "",
-            _parse_content_list(
+            parse_observations(
                 file_contents,
                 "",
             ),
         )
     )
-    assert _validate_conf_content(
+    assert validate_observations_input(
         "",
-        _parse_content_list(
+        parse_observations(
             file_contents,
             "",
         ),
@@ -243,15 +244,18 @@ def test_that_common_observation_error_validation_is_handled(
         else "RESTART = 1; VALUE=1.0; KEY = FOPR;"
     )
     with pytest.raises(ObservationConfigError, match=match):
-        parse_content(
-            f"""
+        validate_observations_input(
+            ".",
+            parse_observations(
+                f"""
                         {obs_type}  FOPR
                         {{
                             {obs_content}
                             {additional}
                         }};
                         """,
-            "",
+                "",
+            ),
         )
 
 
@@ -440,13 +444,15 @@ def test_that_common_observation_error_validation_is_handled(
 )
 def test_that_summary_observation_validation_is_handled(obs_content, match):
     with pytest.raises(ObservationConfigError, match=match):
-        parse_content(obs_content, filename="")
+        validate_observations_input(".", parse_observations(obs_content, filename=""))
 
 
 def test_that_general_observation_without_error_is_invalid():
     with pytest.raises(ObservationConfigError, match="ERROR must also be given"):
-        parse_content(
-            """
+        validate_observations_input(
+            ".",
+            parse_observations(
+                """
             GENERAL_OBSERVATION  obs
             {
                DATA       = RES;
@@ -454,14 +460,17 @@ def test_that_general_observation_without_error_is_invalid():
                VALUE      = 1;
             };
             """,
-            "",
+                "",
+            ),
         )
 
 
 def test_that_general_observation_without_data_is_invalid():
     with pytest.raises(ObservationConfigError, match='Missing item "DATA"'):
-        parse_content(
-            """
+        validate_observations_input(
+            ".",
+            parse_observations(
+                """
             GENERAL_OBSERVATION  obs
             {
                DATE       = 2023-02-01;
@@ -470,7 +479,8 @@ def test_that_general_observation_without_data_is_invalid():
                ERROR_MIN  = 0.1;
             };
             """,
-            "",
+                "",
+            ),
         )
 
 
@@ -480,7 +490,12 @@ def test_that_general_observation_without_data_is_invalid():
 )
 def test_that_unknown_key_is_handled(observation_type):
     with pytest.raises(ObservationConfigError, match="Unknown SMERROR"):
-        parse_content(f"{observation_type} FOPR {{SMERROR=0.1;DATA=key;}};", "")
+        validate_observations_input(
+            ".",
+            parse_observations(
+                f"{observation_type} FOPR {{SMERROR=0.1;DATA=key;}};", ""
+            ),
+        )
 
 
 def test_unexpected_character_handling():
@@ -489,7 +504,7 @@ def test_unexpected_character_handling():
         match=r"Did not expect character: \$ \(on line 4: *ERROR *\$"
         r" 0.20;\). Expected one of {'EQUAL'}",
     ) as err_record:
-        parse_content(
+        parse_observations(
             """
             GENERAL_OBSERVATION GEN_OBS
             {
